multi: Separate unit tests
Move all test code to their own files. Name all test files so that they
can be found by the unittest module. Initiate uninitiated packages.
