+++ pydecred/test_pydecred.py
@@ -1,3 +1,8 @@
+"""
+Copyright (c) 2019, the Decred developers
+See LICENSE for details
+"""
+
 import unittest
 import os
 import json
